Add folders for Exceptions classes
