+++ nvQuickSite/Controllers/IISController.cs
@@ -21,6 +21,7 @@ namespace nvQuickSite.Controllers
     using System.Globalization;
 
     using Microsoft.Web.Administration;
+    using nvQuickSite.Controllers.Exceptions;
 
     /// <summary>
     /// Controls IIS operations.
