+++ nvQuickSite/Exceptions/ReadAndExtractException.cs
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with nvQuickSite.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace nvQuickSite
+namespace nvQuickSite.Exceptions
 {
     using System;
     using System.Runtime.Serialization;
